backend/requirements.txt
fastapi
uvicorn
python-multipart
docling
camelot-py[cv]
paddleocr
paddlepaddle
python-docx
openpyxl
file-saver
python-magic
